+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 
 ### Improvements
 - Removed the PvP balance-loss feature and moved it to the new bounty add-on.
+- `/orders claim` has been changed to `/deliveries`.
 
 ### Fixes